Introduce search functionality in Online shop

+++ src/main/java/org/ogorodnik/shop/web/servlet/ItemsServlet.java
@@ -44,11 +44,20 @@ public class ItemsServlet extends HttpServlet {
             response.sendRedirect("/login");
         }
 
+        String searchItem = request.getParameter("search");
         Map<String, Object> paramsMap = new HashMap<>();
-        try {
-            paramsMap.put("items", itemService.getAll());
-        } catch (SQLException e) {
-            e.printStackTrace();
+        if(null == searchItem || searchItem.isBlank()) {
+            try {
+                paramsMap.put("items", itemService.getAll());
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        } else{
+            try {
+                paramsMap.put("items", itemService.search(searchItem));
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
         }
 
         PageGeneratorCreator pageGeneratorCreator = new PageGeneratorCreator();
@@ -60,15 +69,23 @@ public class ItemsServlet extends HttpServlet {
 
     protected void doPost(HttpServletRequest request,
                           HttpServletResponse response) throws IOException {
-        long id = Long.parseLong(request.getParameter("id"));
+        String searchItem = request.getParameter("search");
+        if(!(null == searchItem) && !searchItem.isBlank()){
+            doGet(request, response);
+        } else {
+            long id = request.getParameter("id") == null ? 0 : Long.parseLong(request.getParameter("id"));
 
-        try {
-            itemService.deleteItem(id);
-        } catch (SQLException e) {
-            e.printStackTrace();
+            if (0 != id) {
+                try {
+                    itemService.deleteItem(id);
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+                doGet(request, response);
+            } else {
+                doGet(request, response);
+            }
         }
-
-        doGet(request, response);
     }
 
     public void setItemService(ItemService itemService) {
